feat(promotions): implement promotion product management with CRUD operations and DTOs

+++ src/main/java/co/hublots/ln_foot/dto/ProductDto.java
@@ -1,6 +1,5 @@
 package co.hublots.ln_foot.dto;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -33,7 +32,7 @@ public class ProductDto {
 
         @NotNull(message = "Price is required")
         @Positive(message = "Price must be positive")
-        private BigDecimal price;
+        private double price;
 
         @NotBlank(message = "Stock quantity is required")
         @Positive(message = "Stock quantity must be positive")
